Get rid of the old versions.
