+++ variables/parse.go
@@ -8,6 +8,10 @@ import (
 	"github.com/pkg/errors"
 )
 
+// ProcessNonConstantVariableFunc is a function which takes in an expression and
+// turns it into a state, target intput and arg index.
+type ProcessNonConstantVariableFunc func(name string, expression string) (value string, argIndex int, err error)
+
 // ParseCommandLineArgs parses a slice of constant build args and returns a new scope.
 func ParseCommandLineArgs(args []string, dotEnvMap map[string]string) (*Scope, error) {
 	ret := NewScope()
